--- .env.example
@@ -1,8 +1,0 @@
-# These are retrieved from your project at app.tina.io
-NEXT_PUBLIC_TINA_CLIENT_ID=***
-TINA_TOKEN=***
-
-# This is set by default CI with Netlify/Vercel/Github, but can be overriden
-NEXT_PUBLIC_TINA_BRANCH=***
-# To see the preview functionality
-VERCEL_ENV=preview
