+++ app/[filename]/page.tsx
@@ -1,0 +1,31 @@
+import React from "react";
+import client from "../../tina/__generated__/client";
+import ClientPage from "./client-page";
+import Layout from "../../components/layout/layout";
+
+export default async function Page({
+  params,
+}: {
+  params: { filename: string[] };
+}) {
+  const data = await client.queries.page({
+    relativePath: `${params.filename}.md`,
+  });
+
+  return (
+    <Layout rawPageData={data}>
+      <ClientPage {...data}></ClientPage>
+    </Layout>
+  );
+}
+
+export async function generateStaticParams() {
+  const pages = await client.queries.pageConnection();
+  return pages.data?.pageConnection.edges.map((edge) => {
+    return {
+      params: {
+        filename: edge.node._sys.filename.split("/"),
+      },
+    };
+  });
+}
