+++ app/layout.tsx
@@ -1,0 +1,48 @@
+import "../styles.css";
+import React from "react";
+import { ThemeProvider } from "../components/theme-provider";
+import { Inter as FontSans } from "next/font/google";
+import { cn } from "../lib/utils";
+import { Metadata } from "next";
+import client from "../tina/__generated__/client";
+
+const fontSans = FontSans({
+  subsets: ["latin"],
+  variable: "--font-sans",
+});
+
+export const metadata: Metadata = {
+  title: "Tina",
+  description: "Tina Cloud Starter",
+};
+
+export default async function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const globalQuery = await client.queries.global({
+    relativePath: "index.json",
+  });
+  const global = globalQuery.data.global;
+
+  return (
+    <html lang="en" suppressHydrationWarning>
+      <body
+        className={cn(
+          "min-h-screen flex flex-col font-sans antialiased",
+          fontSans.variable
+        )}
+      >
+        <ThemeProvider
+          attribute="class"
+          defaultTheme="system"
+          disableTransitionOnChange
+          forcedTheme={global.theme.darkMode}
+        >
+          {children}
+        </ThemeProvider>
+      </body>
+    </html>
+  );
+}
