+++ app/posts/page.tsx
@@ -1,0 +1,18 @@
+// import { InferGetStaticPropsType } from "next";
+import Layout from "../../components/layout/layout";
+import client from "../../tina/__generated__/client";
+import PostsClientPage from "./client-page";
+
+export default async function PostsPage() {
+  const posts = await client.queries.postConnection();
+
+  if (!posts) {
+    return null;
+  }
+
+  return (
+    <Layout rawPageData={posts.data}>
+      <PostsClientPage data={posts.data?.postConnection.edges || []} />
+    </Layout>
+  );
+}
