+++ components/blocks/content.tsx
@@ -1,10 +1,12 @@
+"use client";
 import React from "react";
-import { Container } from "../util/container";
-import { Section } from "../util/section";
+
 import { TinaMarkdown } from "tinacms/dist/rich-text";
 import type { TinaTemplate } from "tinacms";
 import { PageBlocksContent } from "../../tina/__generated__/types";
 import { tinaField } from "tinacms/dist/react";
+import { Container } from "../layout/container";
+import { Section } from "../layout/section";
 
 export const Content = ({ data }: { data: PageBlocksContent }) => {
   return (
