+++ components/blocks/index.tsx
@@ -1,9 +1,9 @@
-import type { Page, PageBlocks } from "../tina/__generated__/types";
-import { Content } from "./blocks/content";
-import { Features } from "./blocks/features";
-import { Hero } from "./blocks/hero";
-import { Testimonial } from "./blocks/testimonial";
 import { tinaField } from "tinacms/dist/react";
+import { Page, PageBlocks } from "../../tina/__generated__/types";
+import { Hero } from "./hero";
+import { Content } from "./content";
+import { Features } from "./features";
+import { Testimonial } from "./testimonial";
 
 export const Blocks = (props: Omit<Page, "id" | "_sys" | "_values">) => {
   return (
@@ -23,10 +23,10 @@ export const Blocks = (props: Omit<Page, "id" | "_sys" | "_values">) => {
 
 const Block = (block: PageBlocks) => {
   switch (block.__typename) {
-    case "PageBlocksContent":
-      return <Content data={block} />;
     case "PageBlocksHero":
       return <Hero data={block} />;
+    case "PageBlocksContent":
+      return <Content data={block} />;
     case "PageBlocksFeatures":
       return <Features data={block} />;
     case "PageBlocksTestimonial":
