+++ components/blocks/testimonial.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import { Container } from "../util/container";
-import { Section } from "../util/section";
 import type { TinaTemplate } from "tinacms";
 import { PageBlocksTestimonial } from "../../tina/__generated__/types";
 import { tinaField } from "tinacms/dist/react";
+import { Section } from "../layout/section";
+import { Container } from "../layout/container";
 
 export const Testimonial = ({ data }: { data: PageBlocksTestimonial }) => {
   return (
