+++ components/layout/container.tsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { cn } from "../../lib/utils";
 
 export const Container = ({
   children,
@@ -23,7 +24,12 @@ export const Container = ({
 
   return (
     <div
-      className={`${widthClass[width]} mx-auto px-6 sm:px-8 ${verticalPadding[size]} ${className}`}
+      className={cn(
+        widthClass[width],
+        `mx-auto px-6 sm:px-8`,
+        verticalPadding[size],
+        className
+      )}
       {...props}
     >
       {children}
