--- components/layout/index.tsx
@@ -1,1 +1,0 @@
-export { useTheme } from "./theme";
