+++ components/layout/section.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import { useTheme } from "../layout";
+import { useLayout } from "../layout/layout-context";
 
 export const Section = ({ children, color = "", className = "" }) => {
-  const theme = useTheme();
+  const { theme } = useLayout();
   const sectionColor = {
     default:
       "text-gray-800 dark:text-gray-50 bg-gradient-to-tl from-gray-50 dark:from-gray-900 via-transparent to-transparent",
