+++ components/nav/header.tsx
@@ -1,0 +1,75 @@
+"use client";
+
+import React from "react";
+import Link from "next/link";
+import { Container } from "../layout/container";
+import { cn } from "../../lib/utils";
+import { tinaField } from "tinacms/dist/react";
+import { Icon } from "../icon";
+import NavItems from "./nav-items";
+import { useLayout } from "../layout/layout-context";
+
+const headerColor = {
+  default:
+    "text-black dark:text-white from-gray-50 to-white dark:from-gray-800 dark:to-gray-900",
+  primary: {
+    blue: "text-white from-blue-300 to-blue-500",
+    teal: "text-white from-teal-400 to-teal-500",
+    green: "text-white from-green-400 to-green-500",
+    red: "text-white from-red-400 to-red-500",
+    pink: "text-white from-pink-400 to-pink-500",
+    purple: "text-white from-purple-400 to-purple-500",
+    orange: "text-white from-orange-400 to-orange-500",
+    yellow: "text-white from-yellow-400 to-yellow-500",
+  },
+};
+
+export default function Header() {
+  const { globalSettings, theme } = useLayout();
+  const header = globalSettings.header;
+
+  const headerColorCss =
+    header.color === "primary"
+      ? headerColor.primary[theme.color]
+      : headerColor.default;
+
+  return (
+    <div
+      className={`relative overflow-hidden bg-gradient-to-b ${headerColorCss}`}
+    >
+      <Container size="custom" className="py-0 relative z-10 max-w-8xl">
+        <div className="flex items-center justify-between gap-6">
+          <h4 className="select-none text-lg font-bold tracking-tight my-4 transition duration-150 ease-out transform">
+            <Link
+              href="/"
+              className="flex gap-1 items-center whitespace-nowrap tracking-[.002em]"
+            >
+              <Icon
+                tinaField={tinaField(header, "icon")}
+                parentColor={header.color}
+                data={{
+                  name: header.icon.name,
+                  color: header.icon.color,
+                  style: header.icon.style,
+                }}
+              />{" "}
+              <span data-tina-field={tinaField(header, "name")}>
+                {header.name}
+              </span>
+            </Link>
+          </h4>
+          <NavItems navs={header.nav} />
+        </div>
+        <div
+          className={cn(
+            `absolute h-1 bg-gradient-to-r from-transparent`,
+            theme.darkMode === "primary"
+              ? `via-white`
+              : `via-black dark:via-white`,
+            "to-transparent bottom-0 left-4 right-4 -z-1 opacity-5"
+          )}
+        />
+      </Container>
+    </div>
+  );
+}
