+++ content/pages/home.md
@@ -73,4 +73,3 @@ blocks:
     color: primary
     _template: testimonial
 ---
-
