+++ content/posts/anotherPost.mdx
@@ -1,5 +1,5 @@
 ---
-title: "Just Another Blog Post "
+title: "Just Another Blog Post"
 heroImg: /uploads/unsplash-lylCw4zcA7I.jpg
 excerpt: >
   Lorem markdownum evinctus ut cape adhaeret gravis licet progenies ut haesit
