+++ content/posts/june/June-Blog-Post.mdx
@@ -1,0 +1,10 @@
+---
+title: June Blog Post
+heroImg: /uploads/llama-2.avif
+excerpt: |
+  Woot asdasdawd
+author: content/authors/pedro.md
+date: 2024-06-17T14:00:00.000Z
+---
+
+Apples
