+++ package.json
@@ -7,11 +7,10 @@
     "build": "tinacms build && next build",
     "start": "tinacms build && next start",
     "export": "npm run build && next export",
-    "lint": "next lint"
+    "lint": "next lint",
+    "dev:build": "next build"
   },
   "devDependencies": {
-    "eslint": "^8",
-    "eslint-config-next": "14.2.3",
     "@svgr/webpack": "^8.1.0",
     "@tinacms/cli": "^1.5.34",
     "@types/js-cookie": "^3.0.0",
@@ -19,6 +18,8 @@
     "@types/react": "^18.3.3",
     "@types/styled-components": "^5.1.15",
     "autoprefixer": "^10.4.0",
+    "eslint": "^8",
+    "eslint-config-next": "14.2.3",
     "postcss": "^8.3.11",
     "postcss-import": "^16.1.0",
     "postcss-nesting": "^12.1.5"
@@ -26,14 +27,18 @@
   "dependencies": {
     "@headlessui/react": "^1.7.19",
     "@tailwindcss/typography": "^0.5.8",
+    "class-variance-authority": "^0.7.0",
     "date-fns": "^3.6.0",
     "next": "^14.2.4",
+    "next-themes": "^0.3.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^5.2.1",
     "styled-jsx": "^5.1.1",
+    "tailwind-merge": "^2.3.0",
     "tailwindcss": "^3.2.4",
     "tinacms": "^1.6.4",
     "typescript": "^5.4.5"
-  }
+  },
+  "packageManager": "pnpm@9.2.0+sha512.98a80fd11c2e7096747762304106432b3ddc67dcf54b5a8c01c93f68a2cd5e05e6821849522a06fb76284d41a2660d5e334f2ee3bbf29183bf2e739b1dafa771"
 }
