--- pages/_app.tsx
@@ -1,7 +1,0 @@
-import "../styles.css";
-
-const App = ({ Component, pageProps }) => {
-  return <Component {...pageProps} />;
-};
-
-export default App;
