+++ tina/collection/page.ts
@@ -39,8 +39,7 @@ const Page: Collection = {
       },
       templates: [
         heroBlockSchema,
-        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-        // @ts-ignore
+        //@ts-ignore
         featureBlockSchema,
         contentBlockSchema,
         testimonialBlockSchema,
